Drop implementation of vec3 to nalgebra crate

+++ src/random.rs
@@ -1,12 +1,12 @@
 use rand::Rng;
 
-pub fn random() -> f64 {
+pub fn random() -> f32 {
   // Generate a random number in [0,1)
   let mut rng = rand::thread_rng();
   rng.gen()
 }
 
-pub fn random_range(min: f64, max: f64) -> f64{
+pub fn random_range(min: f32, max: f32) -> f32{
   // Generate a random number in [min,max)
   min + (max - min) * random()
 }
